break imaginary-specific code out to the message router
